% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDemoStaging.R
\name{loadDemoStaging}
\alias{loadDemoStaging}
\title{Creates randomly-generated data in the staging DB tables, and writes those tables to the staging database via the provided connection object}
\usage{
loadDemoStaging(connection, incidentCount = 1000,
  maxIncidentDate = Sys.Date(), minIncidentDate = function(d) {     d -
  dyears(3) }, seed = NULL)
}
\arguments{
\item{connection}{the DBI connection to use for writing the data}

\item{incidentCount}{the number of Incident records to create and write}

\item{maxIncidentDate}{the latest Incident::IncidentDate to generate}

\item{minIncidentDate}{the earliest Incident::IncidentDate to generate, either as a date object, or a function that takes the maxIncidentDate value and
returns a date object}

\item{seed}{the seed to use for random data generation (set this to the same value for subsequent runs to get the same data generated)}
}
\description{
Creates randomly-generated data in the staging DB tables, and writes those tables to the staging database via the provided connection object
}
